+++ module/motion/QuinticWalk/src/QuinticWalk.hpp
@@ -33,22 +33,22 @@ namespace module::motion {
         ReactionHandle imu_reaction{};
 
         void calculate_joint_goals();
-        [[nodiscard]] float get_time_delta();
+        [[nodiscard]] double get_time_delta();
         [[nodiscard]] std::unique_ptr<message::behaviour::ServoCommands> motion(
-            const std::vector<std::pair<utility::input::ServoID, float>>& joints);
+            const std::vector<std::pair<utility::input::ServoID, double>>& joints);
 
         struct Config {
-            Eigen::Vector3f max_step = Eigen::Vector3f::Zero();
-            float max_step_xy        = 0.0f;
+            Eigen::Vector3d max_step = Eigen::Vector3d::Zero();
+            double max_step_xy       = 0.0f;
 
-            bool imu_active           = true;
-            float imu_pitch_threshold = 0.0f;
-            float imu_roll_threshold  = 0.0f;
+            bool imu_active            = true;
+            double imu_pitch_threshold = 0.0f;
+            double imu_roll_threshold  = 0.0f;
 
             utility::skill::WalkingParameter params{};
 
-            std::map<utility::input::ServoID, float> jointGains{};
-            std::vector<std::pair<utility::input::ServoID, float>> arm_positions{};
+            std::map<utility::input::ServoID, double> jointGains{};
+            std::vector<std::pair<utility::input::ServoID, double>> arm_positions{};
         } normal_config{}, goalie_config{};
 
         static void load_quintic_walk(const ::extension::Configuration& cfg, Config& config);
@@ -56,7 +56,7 @@ namespace module::motion {
         Config& current_config = normal_config;
         bool first_config      = true;
 
-        Eigen::Vector3f current_orders = Eigen::Vector3f::Zero();
+        Eigen::Vector3d current_orders = Eigen::Vector3d::Zero();
         bool is_left_support           = true;
         bool falling                   = false;
         bool first_run                 = true;
